+++ NEWS.md
@@ -1,3 +1,17 @@
+# ggeffects 1.5.1
+
+## General
+
+* Overhaul of the documentation (again), to provide more clarity about the
+  terminology "adjusted predictions", "marginal means" and "marginal effects",
+  and how to calculate each of these quantities using the *ggeffects* package.
+
+* `print_html()` methods were updated to work with the latest release of
+  _tinytable_.
+
+* New `print_md()` method, to print the output as markdown table. This is useful
+  inside RMarkdown or Quarto documents, where the output can be directly included.
+
 # ggeffects 1.5.0
 
 ## New functions